+++ lib/config.ts
@@ -1,5 +1,6 @@
 export const config = {
   charactrAPIUrl: "https://api.charactr.com",
+  charactrAPIUrlWs: "wss://api.charactr.com",
 };
 
 Object.freeze(config);